Refactor: Move tests to separate files
